requirements.txt
selenium
pytrends
pandas
python-dotenv
webdriver-manager

